% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeEnrich.R
\name{cellTypeEnrich}
\alias{cellTypeEnrich}
\title{Cell type enrichment}
\usage{
cellTypeEnrich(
  cells,
  cellTypes,
  mutCells,
  mutCells_cutoff = 5,
  cellType_cutoff = 50
)
}
\arguments{
\item{cells}{barcode of all cells}

\item{cellTypes}{The cell types of each cell}

\item{mutCells}{barcode of mutant cells}

\item{mutCells_cutoff}{This threshold is used to determine how many mutant cells to exceed before cell type
enrichment analysis is performed.}

\item{cellType_cutoff}{This threshold to select cell types with a certain number of cells to consider for inclusion in the analysis}
}
\description{
Cell type enrichment of mutant cells(raw or imputed)
}
