% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutTargetExp.R
\name{mutTargetExp}
\alias{mutTargetExp}
\title{The impact of muttaion in target gene expression (single-cell level or sample-level)}
\usage{
mutTargetExp(SNV.df, SNV.anno, tpm)
}
\arguments{
\item{SNV.df}{Combined mutations for samples (or cells)}

\item{SNV.anno}{mutation annotated file}

\item{tpm}{gene expression matrix}
}
\description{
Cell type enrichment of mutant cells(raw or imputed)
}
