% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImpact.R
\name{plotImpact}
\alias{plotImpact}
\title{To map the effects of mutations on the expression of their target genes, including the spectrum of mutations and the expression of target genes significantly affected by them (up- or down-regulated).}
\usage{
plotImpact(SNV.df, tpm, result, p.cutOff = 0.05)
}
\arguments{
\item{SNV.df}{Combined mutations for samples (or cells)}

\item{tpm}{gene expression matrix}

\item{result}{the comparsion of target gene expression between mutated samples(or cells) and wilde type \code{\link[=obtained from mutTargetExp]{obtained from mutTargetExp()}}}

\item{p.cutOff}{the cutoff of p-value for selected mutations}
}
\description{
SNV Imputation
}
