% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSNVSamples.R
\name{plotSNVSamples}
\alias{plotSNVSamples}
\title{The umap of SNV Imputation}
\usage{
plotSNVSamples(SNV.df, mut.show = NULL)
}
\arguments{
\item{SNV.df}{The combined SNV information from SNVExtract() (data.frame)}

\item{mut.show}{The index of mutations wanted to show (vector)}
}
\description{
SNV Imputation
}
