% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUMAP.R
\name{plotUMAP}
\alias{plotUMAP}
\title{The umap of SNV Imputation}
\usage{
plotUMAP(UMAP.loc, TRS.df)
}
\arguments{
\item{UMAP.loc}{The location of UMAP}

\item{TRS.df}{The result of SNV imputation of SNVImputation()}
}
\description{
SNV Imputation
}
