% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVExtract.R
\name{SNVExtract}
\alias{SNVExtract}
\title{SNV extract}
\usage{
SNVExtract(inputDir)
}
\arguments{
\item{inputDir}{direction of somatic mutations of Monopogen output}
}
\description{
extract SNV from Monopogen outputs
}
