% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVImputation.R
\name{SNVImputation}
\alias{SNVImputation}
\title{SNV Imputation}
\usage{
SNVImputation(
  countMatrix,
  knnGraph = NULL,
  SNVMatrix,
  mutations = row.names(SNVMatrix),
  numk = 30,
  queryCell_cutoff = 5,
  ncors = 10
)
}
\arguments{
\item{countMatrix}{direction of somatic mutations of Monopogen output}

\item{knnGraph}{direction of somatic mutations of Monopogen output}

\item{SNVMatrix}{direction of somatic mutations of Monopogen output}

\item{mutations}{direction of somatic mutations of Monopogen output}

\item{numk}{direction of somatic mutations of Monopogen output}

\item{queryCell_cutoff}{direction of somatic mutations of Monopogen output}

\item{ncors}{direction of somatic mutations of Monopogen output}
}
\description{
SNV Imputation
}
