% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVMatrix.R
\name{SNVMatrix}
\alias{SNVMatrix}
\title{SNV matrix construction for each sample}
\usage{
SNVMatrix(inputDir)
}
\arguments{
\item{inputDir}{direction of somatic mutations of Monopogen output}
}
\description{
contruct mutation matrix for one samples from Monopogen outputs
}
