% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVMatrixExtract.R
\name{SNVMatrixExtract}
\alias{SNVMatrixExtract}
\title{SNV matrix construction for each sample}
\usage{
SNVMatrixExtract(inputDirs)
}
\arguments{
\item{inputDirs}{direction of somatic mutations of Monopogen output(named strings or named strings vector)}
}
\description{
contruct mutation matrix for one samples from Monopogen outputs
}
