% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAPImputation.R
\name{UMAPImputation}
\alias{UMAPImputation}
\title{SNV Imputation umap}
\usage{
UMAPImputation(UMAP.loc, TRS.df)
}
\arguments{
\item{UMAP.loc}{The location of UMAP}

\item{TRS.df}{The result of SNV imputation of SNVImputation}
}
\description{
SNV Imputation
}
