% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNVExtract.R
\name{SNVExtract}
\alias{SNVExtract}
\title{SNV extract}
\usage{
SNVExtract(inputDirs)
}
\arguments{
\item{inputDirs}{direction of somatic mutations of Monopogen output(named strings or named strings vector)}
}
\description{
extract SNV from Monopogen outputs
}
